requirements.txt
# 🚀 CitySpark Complete Project - Python Dependencies
# Core Requirements
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
# psycopg2==2.9.9

# Web Scraping & APIs
requests==2.31.0
beautifulsoup4==4.12.2
github3.py==4.0.1
aiohttp==3.9.1

# AI/ML Core
# torch==2.1.1
# torchvision==0.16.1
transformers==4.36.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3

# Computer Vision & Image Processing
opencv-python==4.8.1.78
pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0

# NLP & Text Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
gensim==4.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Background Tasks & Queues
celery==5.3.4
redis==5.0.1

# File Processing
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# Data Validation & Serialization
marshmallow==3.20.1
jsonschema==4.20.0

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# API Documentation
fastapi-users==12.1.2
fastapi-pagination==0.12.13

# WebSocket Support
websockets==12.0

# HTTP Client
httpx==0.25.2

# Cache
aioredis==2.0.1

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0
python-dateutil==2.8.2
pytz==2023.3

# LuminScript Support
jinja2==3.1.2
markupsafe==2.1.3

# Development Server
watchdog==3.0.0